MVAPICH2 2.3b release
git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@8614 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ osu_benchmarks/mpi/collective/osu_coll.c
@@ -225,7 +225,7 @@ set_num_probes (int value)
 }
 
 static int
-set_max_memlimit (int value)
+set_max_memlimit (long long value)
 {
     options.max_mem_limit = value;
